pyproject.toml
[project]
name = "negotiation-delegation"
version = "0.1.0"
description = "AI delegation for real-world negotiation scenarios. SPI 352 (Fall 2025) final project."
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.34.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

